change type and add logging

+++ speaker-recognition/src/main/java/com/ericsson/ai/speaker/dao/impl/SpeakerProfileDaoImpl.java
@@ -1,6 +1,6 @@
 package com.ericsson.ai.speaker.dao.impl;
 
-import java.util.UUID;
+import java.util.logging.Logger;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
@@ -14,26 +14,43 @@ import com.ericsson.ai.speaker.domain.SpeakerProfile;
 @Repository
 public class SpeakerProfileDaoImpl implements SpeakerProfileDao
 {
+    private static Logger _logger = Logger.getLogger(SpeakerProfileDaoImpl.class.getName());
+
     @PersistenceContext
     private EntityManager _entityManager;
 
     @Override
-    public SpeakerProfile getSpeakerProfile(UUID pProfileId)
+    public SpeakerProfile getSpeakerProfile(String pProfileId)
     {
-       return _entityManager.find(SpeakerProfile.class, pProfileId);
+       SpeakerProfile speakerProfile = _entityManager.find(SpeakerProfile.class, pProfileId);
+       if(speakerProfile == null)
+       {
+           _logger.warning("The speaker profile does not exist for ID: " + pProfileId);
+       }
+       return speakerProfile;
     }
 
     @Override
     @Transactional
     public void save(SpeakerProfile pSpeakerProfile)
     {
+        if(pSpeakerProfile == null)
+        {
+            _logger.warning("The try to save a NULL speaker profile");
+            return;
+        }
         _entityManager.persist(pSpeakerProfile);
     }
 
     @Override
-    @Transactional
+    @Transactional 
     public void delete(SpeakerProfile pSpeakerProfile)
     {
+        if(pSpeakerProfile == null)
+        {
+            _logger.warning("The try to delete a NULL speaker profile");
+            return;
+        }
         _entityManager.remove(pSpeakerProfile);
     }
